remove functions from script and move them into module utils
